SAR Entity
This is a VHDL model of a general successive approximation register type ADC core. The design is done for an eight bit architecture. The implementation of the VHDL model is using behavioural modelling based on a Mealy state machine. It incorporates 10 states as eight states for each bit in the output and one state each for the starting the conversion and concluding the conversion which includes updating the 'Done' flag bit and updating the output register with the result.