SAR Test Bench
This is the test bench to test the designed SAR entity. SAR entity is a finite state machine driven by a periodic clock signal. 

The test bench generates all possible comparator input sequences for an 8 bit SAR type ADC. SAR entity takes one comparator input bit as an input in every clock cycle that it determines an output bit. Thus, including the start state and the end state, there are ten states altogether in the design. 

The test bench models a reset signal and a clock signal as separate VHDL processes ( rst and clock respectively ).

The input signals to the SAR entity and the output signals generated from the SAR entity are handled in the process simulation.

Since the range of test input sequences is as high as 256, the test bench uses a text file output which writes the input comparator bit sequence and the 8 bit SAR output as a comma separated pair in each line. 

The test bench also uses the assert and report statements in VHDL to check if the input comparator bit sequence and the SAR output do not match and report an error to the simulation console. Both these features are implemented in a separate process called error_reporting which is triggered every time the 'Done' bit goes high.